Move unpack target helpers into assign/del module
