+++ src/transform/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod context;
 pub(crate) mod expr;
 pub(crate) mod rewrite_assert;
+pub(crate) mod rewrite_assign_del;
 pub(crate) mod rewrite_class_def;
 pub(crate) mod rewrite_decorator;
 pub(crate) mod rewrite_expr_to_stmt;
